+++ src/array/reducer/add-control.ts
@@ -14,14 +14,14 @@ export function addControlReducer<TValue>(
     return childReducer(state, action);
   }
 
-  const index = action.payload.index === undefined ? state.controls.length : action.payload.index;
+  const index = action.index === undefined ? state.controls.length : action.index;
 
   if (index > state.controls.length || index < 0) {
     throw new Error(`Index ${index} is out of bounds for array '${state.id}' with length ${state.controls.length}!`); // `;
   }
 
   let controls = [...state.controls];
-  controls.splice(index, 0, createChildState(`${state.id}.${index}`, action.payload.value) as InferredControlState<TValue>);
+  controls.splice(index, 0, createChildState(`${state.id}.${index}`, action.value) as InferredControlState<TValue>);
   controls = controls.map((c, i) => ({ ...(c as any), id: `${state.id}.${i}` }));
 
   return computeArrayState(