+++ src/control/reducer/clear-async-error.ts
@@ -10,11 +10,11 @@ export function clearAsyncErrorReducer<TValue extends FormControlValueTypes>(
     return state;
   }
 
-  if (state.pendingValidations.indexOf(action.payload.name) < 0) {
+  if (state.pendingValidations.indexOf(action.name) < 0) {
     return state;
   }
 
-  const name = `$${action.payload.name}`;
+  const name = `$${action.name}`;
 
   let errors = state.errors;
 
@@ -23,7 +23,7 @@ export function clearAsyncErrorReducer<TValue extends FormControlValueTypes>(
     delete errors[name];
   }
 
-  const pendingValidations = state.pendingValidations.filter(v => v !== action.payload.name);
+  const pendingValidations = state.pendingValidations.filter(v => v !== action.name);
   const isValid = isEmpty(errors);
 
   return {