+++ src/control/reducer/start-async-validation.ts
@@ -9,13 +9,13 @@ export function startAsyncValidationReducer<TValue extends FormControlValueTypes
     return state;
   }
 
-  if (state.pendingValidations.indexOf(action.payload.name) >= 0) {
+  if (state.pendingValidations.indexOf(action.name) >= 0) {
     return state;
   }
 
   return {
     ...state,
-    pendingValidations: [...state.pendingValidations, action.payload.name],
+    pendingValidations: [...state.pendingValidations, action.name],
     isValidationPending: true,
   };
 }